Ready to submit
I have added a couple of additional features which were not required. There by a middleware is checking if the foreign key actually exists in the database. 

The test tries to add the property Ids z0123456-78ab-cdef-9012-3456789abcdef and z9876543-21fe-dcba-1098-76543210dcba as a new booking and review is being created and expect a response 201 for this. As I don’t believe this is good practice the middleware returns a code 400 for bad request.

Additionally, it is not required that the email is unique. I have however made the email unique as I see it as good practice. It is commonly used that a link can be send to a user’s email for resetting a password. This is however only possible if the email is unique. I have therefore also implemented a middleware returning a response 400 if this is the case. In the test the emails johndoe@example.com and johndoe@email.com is being added for creating a new user and host. These emails are however also in the provided seeded data. As the test expect a response 201 but gets a response 400 these are failing. 

I have however decided to leave in the implemented features as I believe this is better practice.